% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_mean}
\alias{gen_data_mean}
\title{Generate Mean Values for Various Metrics}
\usage{
gen_data_mean(data)
}
\arguments{
\item{data}{A data frame containing the raw data.}
}
\value{
A data frame with mean values for various metrics.
}
\description{
This function calculates mean values for various metrics grouped by source, time, gender, and age.
}
