% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_mean_weighted}
\alias{gen_data_mean_weighted}
\title{Generate Weighted Mean Values Over Years}
\usage{
gen_data_mean_weighted(data_mean)
}
\arguments{
\item{data_mean}{A data frame containing mean values for various metrics.}
}
\value{
A data frame with weighted mean values for various metrics over years.
}
\description{
This function calculates weighted mean values for various metrics over years.
}
