% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_mean_weighted_rf_wide}
\alias{gen_data_mean_weighted_rf_wide}
\title{Calculate Differences for Various Metrics}
\usage{
gen_data_mean_weighted_rf_wide(data_mean_weighted)
}
\arguments{
\item{data_mean_weighted}{A data frame containing weighted mean values for various metrics.}
}
\value{
A data frame with differences between intervention and baseline values for various metrics.
}
\description{
This function calculates the differences between intervention and baseline values for various metrics.
}
