% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_weighted_burden}
\alias{gen_data_weighted_burden}
\title{Calculate Differences for Burden of Disease}
\usage{
gen_data_weighted_burden(data_weighted)
}
\arguments{
\item{data_weighted}{A data frame containing weighted mean values for various
metrics.}
}
\value{
A data frame with differences between intervention and baseline
values for burden of disease.
}
\description{
This function calculates the differences between intervention and baseline
values for burden of disease.
}
