% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_weighted_burden_spline}
\alias{gen_data_weighted_burden_spline}
\title{Perform data smoothing}
\usage{
gen_data_weighted_burden_spline(data_weighted_burden)
}
\arguments{
\item{data_weighted_burden}{A data frame containing weighted values for burden of disease.}
}
\value{
A data frame with spline smoothing applied for burden of disease.
}
\description{
This function performs data smoothing.
}
