% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_weighted_ds_cumdiff}
\alias{gen_data_weighted_ds_cumdiff}
\title{Calculate Cumulative Differences for Incidences}
\usage{
gen_data_weighted_ds_cumdiff(data_weighted)
}
\arguments{
\item{data_weighted}{A data frame containing weighted mean values for various
metrics.}
}
\value{
A data frame with differences between intervention and baseline
values for incidences.
}
\description{
This function calculates the cumulative differences between intervention and
baseline values for incidences.
}
