% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_weighted_rf}
\alias{gen_data_weighted_rf}
\title{Calculate Differences for Risk Factors}
\usage{
gen_data_weighted_rf(data_weighted, configname = "default")
}
\arguments{
\item{data_weighted}{A data frame containing weighted mean values for various
metrics.}

\item{configname}{The name of the configuration file to use (e.g., "default", "development", "production", "testing").}
}
\value{
A data frame with differences between intervention and baseline
values for risk factors.
}
\description{
This function calculates the differences between intervention and baseline
values for risk factors.
}
