% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskfactors.R
\name{inc_diff}
\alias{inc_diff}
\title{Plot of Incidence Difference}
\usage{
inc_diff(inc, data_mean_weighted_inc_wide)
}
\arguments{
\item{inc}{A character string specifying the incidence to plot.
Options are: "asthma", "ckd", "diabetes", "ischemia", "stroke".}

\item{data_mean_weighted_inc_wide}{A data frame containing the weighted mean values of incidences.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Creates a line plot showing the reduction in a specified incidence number over time.
}
