% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-plots.R
\name{life_exp}
\alias{life_exp}
\title{Plot of Life Expectancy under Intervention}
\usage{
life_exp(diff, data_ple_wide)
}
\arguments{
\item{diff}{A character string specifying the life expectancy to plot.}

\item{data_ple_wide}{A data frame containing the life expectancy.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Creates a line plot showing the increase in life expectancy under
intervention over time.
}
