% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_config.R
\name{load_config}
\alias{load_config}
\title{Load configuration file}
\usage{
load_config(config_name = "default")
}
\arguments{
\item{config_name}{The name of the configuration environment to load (e.g.,
"default", "development", "production", "testing").}
}
\value{
A list containing the configuration data.
}
\description{
This function loads the configurations from the config.yml file located
in the inst/config directory using the \code{config} package.
}
