% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskfactors.R
\name{riskfactors_diff}
\alias{riskfactors_diff}
\title{Plot of Difference in Risk Factor}
\usage{
riskfactors_diff(riskft_diff, data_mean_weighted_rf_wide)
}
\arguments{
\item{riskft_diff}{A character string specifying the difference in risk factor to plot.
Options are: "bmi", "ei", "obesity", "sodium".}

\item{data_mean_weighted_rf_wide}{A data frame containing the weighted mean values of risk factors.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Creates a line plot showing the reduction in a specified risk factor under intervention over time.
}
