% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-graphics.R
\name{scale_fun}
\alias{scale_fun}
\title{Format numbers to 2 decimal places}
\usage{
scale_fun(x)
}
\arguments{
\item{x}{Numeric vector to be formatted.}
}
\value{
A character vector with numbers formatted to 2 decimal places.
}
\description{
This function formats a numeric vector to display numbers with 2 decimal
places.
}
\examples{
x <- c(9.3475, 0.1974)
scale_fun(x)

}
