% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_graphics.R
\name{scaleFUN}
\alias{scaleFUN}
\title{Format numbers to 2 decimal places}
\usage{
scaleFUN(x)
}
\arguments{
\item{x}{Numeric vector to be formatted.}
}
\value{
A character vector with numbers formatted to 2 decimal places.
}
\description{
This function formats a numeric vector to display numbers with 2 decimal places.
}
\examples{
x <- c(9.3475, 0.1974)
scaleFUN(x)

}
