% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-graphics.R
\name{suppressWarnings}
\alias{suppressWarnings}
\title{Suppress warnings}
\usage{
suppressWarnings()
}
\description{
This function suppresses warnings. It sets the warning option to -1.
Use \code{options(warn=0)} to reset the warning options to default.
}
\examples{
\dontrun{
# Suppress warnings
suppressWarnings()

# Reset warnings to default
options(warn=0)
}

}
