% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-plots.R
\name{burden_disease}
\alias{burden_disease}
\title{Plot of Burden of Disease}
\usage{
burden_disease(burden, data_mean_weighted_burden_wide)
}
\arguments{
\item{burden}{A character string specifying the burden of disease to plot.
Options are: "daly", "dalycum", "yld", "yll".}

\item{data_mean_weighted_burden_wide}{A data frame containing the weighted mean values of burden.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Creates a line plot showing the reduction in a specified burden of disease over time.
}
