% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{DataProcessing}
\alias{DataProcessing}
\title{Data Processing}
\description{
This file contains a set of functions designed to work together for processing the data.
Below is a description of how to use these functions in sequence.
}
\details{
\subsection{Step-by-Step Usage:}{
\enumerate{
\item \strong{Read the data}: This function reads the data from the location specified \code{data <-  readRDS("data.rds")}.
\item \strong{\code{gen_data_mean}}: Calculates weighted mean values for various metrics over years \code{data_weighted <- gen_data_mean(data)}.
\item \strong{\code{gen_data_weighted_rf}}: Calculates the differences between intervention and baseline values for risk factors \code{data_weighted_rf_wide_collapse <- gen_data_weighted_rf(data_weighted)}.
\item \strong{\code{gen_data_weighted_ds}}: Calculates the differences between intervention and baseline values for incidences \code{data_weighted_ds_wide_collapse <- gen_data_weighted_ds(data_weighted)}.
\item \strong{\code{gen_data_weighted_burden}}: Calculates the differences between intervention and baseline values for burden of disease \code{data_weighted_burden_wide_collapse <- gen_data_weighted_burden(data_weighted)}.
\item \strong{\code{gen_data_weighted_burden_spline}}: Performs data smoothing for burden of disease, when necessary. For instance, with only a few simulations, there can be positive values in difference in burden of disease \code{data_weighted_burden_spline <- gen_data_weighted_burden_spline(data_weighted_burden_wide_collapse)}.
\item \strong{\code{gen_data_le}}: Calculates life expectancy for various age and groups \code{data_ple_wide <- gen_data_le(data_weighted)}.
}
}

\subsection{Examples}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Example of using all functions together
data <- readRDS("data.rds")
data_weighted <- gen_data_mean(data)
data_weighted_rf_wide_collapse <- gen_data_weighted_rf(data_weighted)
data_weighted_ds_wide_collapse <- gen_data_weighted_ds(data_weighted)
data_weighted_burden_wide_collapse <- gen_data_weighted_burden(data_weighted)
data_weighted_burden_spline <- gen_data_weighted_burden_spline(data_weighted_burden_wide_collapse)
data_ple_wide <- gen_data_le(data_weighted)
}\if{html}{\out{</div>}}
}
}
