% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{gen_data_weighted_burden_spline}
\alias{gen_data_weighted_burden_spline}
\title{Perform data smoothing}
\usage{
gen_data_weighted_burden_spline(
  data_weighted_burden_wide_collapse,
  configname = "default"
)
}
\arguments{
\item{data_weighted_burden_wide_collapse}{A data frame containing weighted values for burden of disease.}

\item{configname}{The name of the configuration file to use (e.g., "default", "development", "production", "testing").}
}
\value{
A data frame with spline smoothing applied for burden of disease.
}
\description{
This function performs data smoothing for burden of disease, when necessary. For instance, with only a few simulations, there can be positive values in difference in burden of disease.
}
