% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-plots.R
\name{inc_cum}
\alias{inc_cum}
\title{Plot of Cumulative Incidence Difference}
\usage{
inc_cum(
  inc,
  data_weighted_ds_wide_collapse,
  scale_y_continuous_limits = NULL,
  scale_y_continuous_breaks = ggplot2::waiver(),
  scale_y_continuous_labels = ggplot2::waiver()
)
}
\arguments{
\item{inc}{A character string specifying the incidence to plot.
Options: "asthma", "ckd", "diabetes", "ischemia", "stroke".}

\item{data_weighted_ds_wide_collapse}{A data frame with differences between
intervention and baseline values for incidences.}

\item{scale_y_continuous_limits}{A numeric vector specifying the limits of
the scales for continuous y aesthetics.}

\item{scale_y_continuous_breaks}{A numeric vector specifying the breaks of
the scales for continuous y aesthetics.}

\item{scale_y_continuous_labels}{A numeric vector specifying the labels of
the scales for continuous y aesthetics.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Creates a line plot showing the cumulative reduction in a specified incidence
number over time.
}
