% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-plots.R
\name{inc_diff}
\alias{inc_diff}
\title{Plot of Incidence Difference}
\usage{
inc_diff(inc, data_weighted_ds_wide_diff)
}
\arguments{
\item{inc}{A character string specifying the incidence to plot.
Options: "asthma", "ckd", "diabetes", "ischemia", "stroke".}

\item{data_weighted_ds_wide_diff}{A data frame containing the weighted mean
values of differences of incidences.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Creates a line plot showing the reduction in a specified incidence number
over time.
}
