% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-plots.R
\name{riskfactors}
\alias{riskfactors}
\title{Plot Risk Factors Over Time}
\usage{
riskfactors(riskft, data_mean_weighted)
}
\arguments{
\item{riskft}{A character string specifying the risk factor to plot.
Options are: "bmi", "ei", "fat", "obese", "protein", "sodium".}

\item{data_mean_weighted}{A data frame with weighted mean values for various metrics.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Generates a line plot of a specified risk factor over time, grouped by source.
}
