% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-plots.R
\name{riskfactors_diff}
\alias{riskfactors_diff}
\title{Plot of Difference in Risk Factor}
\usage{
riskfactors_diff(
  riskft_diff,
  data_weighted_rf_wide_collapse,
  scale_y_continuous_limits = NULL,
  scale_y_continuous_breaks = ggplot2::waiver(),
  scale_y_continuous_labels = ggplot2::waiver()
)
}
\arguments{
\item{riskft_diff}{A character string specifying the difference in risk
factor to plot.
Options: "bmi", "ei", "obesity", "sodium".}

\item{data_weighted_rf_wide_collapse}{A data frame with differences between
intervention and baseline values for risk factors.}

\item{scale_y_continuous_limits}{A numeric vector specifying the limits of
the scales for continuous y aesthetics.}

\item{scale_y_continuous_breaks}{A numeric vector specifying the breaks of
the scales for continuous y aesthetics.}

\item{scale_y_continuous_labels}{A numeric vector specifying the labels of
the scales for continuous y aesthetics.}
}
\value{
A ggplot object representing the specified plot.
}
\description{
Creates a line plot showing the reduction in a specified risk factor under
intervention over time.
}
